pyproject.toml
[tool.poetry]
name = "strawberry-graphql-1504"
version = "0.1.0"
description = ""
authors = ["BB <boyan@typed.wtf>"]

[tool.poetry.dependencies]
python = "^3.7"
strawberry-graphql = {version = "^0.93.6", extras = ["debug-server"]}
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
